Remove the mode to run in the browser
