+++ section-report/package.json
@@ -3,12 +3,11 @@
   "version": "1.0.0-sectionreport.1",
   "private": true,
   "scripts": {
-    "serve": "cross-env VUE_APP_HANDLERS=browser vue-cli-service serve",
-    "build": "cross-env VUE_APP_HANDLERS=browser vue-cli-service build",
+    "build": "vue-cli-service build",
     "test:unit": "vue-cli-service test:unit",
     "lint": "vue-cli-service lint",
-    "electron:build": "cross-env VUE_APP_HANDLERS=electron vue-cli-service electron:build",
-    "electron:serve": "cross-env VUE_APP_HANDLERS=electron vue-cli-service electron:serve",
+    "electron:build": "vue-cli-service electron:build",
+    "electron:serve": "vue-cli-service electron:serve",
     "i18n:report": "vue-cli-service i18n:report --src \"./src/**/*.?(js|vue)\" --locales \"./src/locales/**/*.json\"",
     "postinstall": "electron-builder install-app-deps",
     "postuninstall": "electron-builder install-app-deps",
