+++ section-report/public/index.html
@@ -4,8 +4,7 @@
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
-    <link rel="icon" href="<%= BASE_URL %>favicon.ico">
-    <script src="handlers/<%= VUE_APP_HANDLERS %>.js"></script>
+    <link rel="icon" href="favicon.ico">
     <title>Thinreports Editor</title>
   </head>
   <body>
