--- section-report/src/@types/window.d.ts
@@ -1,7 +1,0 @@
-import { Handlers } from '@/types';
-
-declare global {
-  interface Window {
-    handlers: Handlers;
-  }
-}
