--- section-report/src/handlers.ts
@@ -1,1 +1,0 @@
-export default window.handlers;
